+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
  * Adds Arel predications for `ANY` and `ALL` - Dan McClain
  * Fixes errors with has_and_belongs_to_many associations - Jacob Swanner
+ * Adds with.recursive for recursive CTEs - Cody Cutrer
 
 ## 2.1.3